# Make calls from Julia to GAP faster