--- a/test/basics.jl
+++ b/test/basics.jl
@@ -70,6 +70,15 @@ end
     @test GAP.julia_to_gap([1, 2, 3, 4, 5, 6]) == f(1, 2, 3, 4, 5, 6)
     @test GAP.julia_to_gap([1, 2, 3, 4, 5, 6, 7]) == f(1, 2, 3, 4, 5, 6, 7)
 
+    @test [] == convert(Vector{String}, f())
+    @test ["1"] == convert(Vector{String}, f("1"))
+    @test ["1", "2"] == convert(Vector{String}, f("1", "2"))
+    @test ["1", "2", "3"] == convert(Vector{String}, f("1", "2", "3"))
+    @test ["1", "2", "3", "4"] == convert(Vector{String}, f("1", "2", "3", "4"))
+    @test ["1", "2", "3", "4", "5"] == convert(Vector{String}, f("1", "2", "3", "4", "5"))
+    @test ["1", "2", "3", "4", "5", "6"] == convert(Vector{String}, f("1", "2", "3", "4", "5", "6"))
+    @test ["1", "2", "3", "4", "5", "6", "7"] == convert(Vector{String}, f("1", "2", "3", "4", "5", "6", "7"))
+
     # check to see if a non-basic object (here: a tuple) can be
     # passed and then extracted again
     @test f((1, 2, 3))[1] == (1, 2, 3)