**UPDATE: once again rebased and timings updated; also added more extensive GAP microbenchmarking results to show how close we are to matching the GAP interpreter.**

... by avoiding conversion calls for immediate integers, immediate FFE and
GapObj; and also by directly ccalling into the GAP function handlers.

Before this patch:

    julia> @btime test_simple()
      25.844 ms (0 allocations: 0 bytes)

    julia> @btime test_isabelian_1()
      195.323 ms (1000000 allocations: 15.26 MiB)

    julia> @btime test_isabelian_2()
      159.374 ms (1000000 allocations: 15.26 MiB)

After this patch (note the avoided allocations):

    julia> @btime test_simple()
      25.546 ms (0 allocations: 0 bytes)

    julia> @btime test_isabelian_1()
      53.635 ms (0 allocations: 0 bytes)

    julia> @btime test_isabelian_2()
      25.289 ms (0 allocations: 0 bytes)

For comparison, here are timings for roughly equivalent GAP functions;

    julia> @btime GAP.Globals.test_simple()
      5.134 ms (0 allocations: 0 bytes)

    julia> @btime GAP.Globals.test_isabelian_1()
      20.187 ms (0 allocations: 0 bytes)

    julia> @btime GAP.Globals.test_isabelian_2()
      18.176 ms (0 allocations: 0 bytes)

Here is the Julia microbenchmark:
```julia
    using GAP, BenchmarkTools
    function test_simple()
      for i = 1:1000000
        x = GAP.Globals.IsAbelian
      end
    end

    G = GAP.Globals.CyclicGroup(10)

    function test_isabelian_1()
      x = true
      for i = 1:1000000
        x = x && GAP.Globals.IsAbelian(G)
      end
    end

    function test_isabelian_2()
      f = GAP.Globals.IsAbelian
      x = true
      for i = 1:1000000
        x = x && f(G)
      end
    end
```
And here is the GAP version (can be entered using `GAP.prompt()`)
```gap
    test_simple := function()
      local x, i;
      for i in [1.. 1000000] do
        x := IsAbelian;
      od;
    end;

    G := CyclicGroup(10);

    test_isabelian_1 := function()
      local x, i;
      x := true;
      for i in [1.. 1000000] do
        x := x and IsAbelian(G);
      od;
    end;

    test_isabelian_2 := function()
      local f, x, i;
      f := IsAbelian;
      x := true;
      for i in [1.. 1000000] do
        x := x and f(G);
      od;
    end;
```
Resolves #486

TODO:
- [x] add tests that pass in/out various types combos
- [x] add tests that verify that calls with 0-6 args work fine
- [ ] in threaded mode, we ought to call `BeginGapSync` and `EndGapSync` (and then we also need to worry about catching errors, so that we can call `EndGapSync` regardless)