@fingolfin 2020-07-21
The test failures saying `Error, UndefVarError: T_FUNCTION not defined` are caused by the fact that already `GAP_Initialize` ends up processing the `tst/testall.g` file... We somehow need to insert code that is run after the kernel has fully initialized but before GAP's `init.g` processes command line arguments. I guess I could do this by inserting another `-c BLA` into the command line. But on the long run, I think the libgap API needs to be improved for this kinda of stuff.